requirements.txt
fastapi
uvicorn
pydantic
pandas
numpy
scikit-learn
matplotlib
seaborn
plotly
streamlit
httpx
openai
pytest
black
isort
python-multipart
xgboost
langchain-community
langchain-core
langchain-experimental
langchainhub
langchain-google-genai
langgraph